Renamed self_log to selflog and commented.
For some reason when I returned to the code I found self_log made it
hard to read these lines.  I think it is because of adjacency to a
method call on a _foo method coupled with the fact self is usually
immediately followed by a dot.
So I read it as self.log._foo()  and thought what on Earth?

Removing the underscore stopped me going cross-eyed somehow.